Add drivers/* into the empty baseline.
Some sources still left out, as they would require too much stubbing.

+++ drivers/source/CAN.cpp
@@ -31,7 +31,7 @@ CAN::CAN(PinName rd, PinName td) : _can(), _irq()
     }
 
     can_init(&_can, rd, td);
-    can_irq_init(&_can, (&CAN::_irq_handler), (uint32_t)this);
+    can_irq_init(&_can, (&CAN::_irq_handler), (uint32_t)(intptr_t)this);
 }
 
 CAN::CAN(PinName rd, PinName td, int hz) : _can(), _irq()
@@ -43,7 +43,7 @@ CAN::CAN(PinName rd, PinName td, int hz) : _can(), _irq()
     }
 
     can_init_freq(&_can, rd, td, hz);
-    can_irq_init(&_can, (&CAN::_irq_handler), (uint32_t)this);
+    can_irq_init(&_can, (&CAN::_irq_handler), (uint32_t)(intptr_t)this);
 }
 
 CAN::~CAN()
